+++ Arduino/src/output/LedPwm.h
@@ -5,11 +5,12 @@
 #include "../interface/Light.h"
 #include "Led.h"
 
+/** Class to manage a led with support for pwm */
 class LedPwm :  public interface::LightPwm, public Led {
 public:
     LedPwm(int);
     LedPwm(int, int);
-    void setIntensity(int);
+    void setIntensity(uint8_t);
     void switchOn();
     void switchOff();
 private:
